Changed demo 7 to a far more practical use case
If you need to use the mouse to alter the funcitonality of the demo then you've failed keyboard accessibliity.

This new demo temporarily presents a more valid use-case. It represents an author who is waiting for content to load.

The author does not want the user to be able to escape the focus-trap until the content has finished loading.