Add Cypress e2e tests against the demos (#144)
Since 'the demo is also the test' in this repo, let's use the
advent of Cypress (which is able to effectively test things
related to focus, unlike unit tests with JSDom, which doesn't
fully support focus events) to automate running all the demos
and exercise the focus-trap API.

NOTE: `demo/index.html` is UNCHANGED in functionality. I just
added a wrapper div with ID to each demo to make it easier
to find each demo when running Cypress tests.

The framework is in place, and the first demo is tested. Other
demos are TODO. I'm hoping to get community help to complete them.
See #143