Always call onPostDeactivate even if focus not returned
This is consistent with always calling onPostActivate after
onActivate.

I enhanced the Animated Trigger demo to show that
onPostDeactivate is always called even if `returnFocus`
is not set.

And I added visible (i.e. on-screen) messages for
the Animated Dialog and Animated Trigger demos
instead of logging to the console. This way, I could
add a test for the Tigger where focus is not returned
but onPostDeactivate is still called.