Always call onPostDeactivate even if focus not returned (#433)
* Always call onPostDeactivate even if focus not returned

This is consistent with always calling onPostActivate after
onActivate.

I enhanced the Animated Trigger demo to show that
onPostDeactivate is always called even if `returnFocus`
is not set.

And I added visible (i.e. on-screen) messages for
the Animated Dialog and Animated Trigger demos
instead of logging to the console. This way, I could
add a test for the Tigger where focus is not returned
but onPostDeactivate is still called.

* Replace HTML code with character

Co-authored-by: Daniel Tonon <Dan503@users.noreply.github.com>

* Postdeactivate PR suggestions (#434)

Co-authored-by: Daniel Tonon <dan503.npm@gmail.com>
Co-authored-by: Stefan Cameron <stefan@stefcameron.com>

* Update demo bundle

Co-authored-by: Daniel Tonon <Dan503@users.noreply.github.com>
Co-authored-by: Daniel Tonon <dan503.npm@gmail.com>