Current demo was impossible to debug index.js, now it is (#522)
Instead of pre-compiing the CJS bundle and then using that in the
demo (which makes `index.js` completely un-debuggable), we just
include `index.js` as source in the demo bundle. Now it's totally
debuggable.

Also remove the unnecessary `--` to pass parameters in `yarn start`
as this is not needed with yarn (and will later be passed to the
underlying script, whcih would break things).