[#612] Add new isKeyForward and isKeyBackward options
Fixes #612

Using these new options makes it possible to have focus move within
the trap using keys other than the default `tab` and `shift+tab`.

Note that using these options doesn't mean the tab key is automatically
disabled. The user needs to handle it themselves to either suppress it
or make it do something else. focus-trap can't know what it should do,
so it leaves it alone, which means it's still functional to a certain
extent (i.e. will move focus forward/backward until it gets to a
container edge, and then it will stop, probably because it tries
to escape the trap at this point, and focus-trap brings it back in
to its last-focused-node).