Looks like now, when referring to a tag created with `.as(...)`, Cypress re-executes the statement prior that was used to get the element that was tagged instead of just referencing the original element.
And since the 'allow outside click' demo used to change the label
of the 'activate trap' button to 'deactivate trap' upon trap
activation, attempting to address the tag later was failing the
tests because the 'activate trap' button couldn't be found.

So now the button's label is always 'activate/deactivate trap'.