[#893] Fix incorrect returnFocusOnDeactivate=true behavior
Fixes #893

If `clickOutsideDeactivates=true` along with the return focus option being
`true` and the outside click is on a focusable node, focus was returning to
that node instead of the node originally focused prior to activation.

Now the `returnFocusOnDeactivate` option does exactly what it says it will
do instead of the seemingly well-intentioned (based on the comment that was
removed) behavior.