Initial test setup when focused element in trap is removed
If the element that is currently focused is removed from the DOM, the
focus is moved either to the document's body element or becomes null
Source: https://amberwilson.co.uk/blog/where-did-the-focus-go/

When using focus trap it would be appropriate to keep focus inside the
trap.

This adds a Cypress test to show whether this works or not.