Pull focus back into DOM when focused element is removed
If the currently focused element is removed then browsers will reset
focus to window.body. See
https://allyjs.io/tutorials/mutating-active-element.html

This can cause assistive technologies to start from the beginning of a
document even a focus trap is set.

Use MutationObserver (if available) to monitor trap content. If the most
recent element to receive focus is removed then reset focus to the 
initial element in the trap.