Fix demo-bundle issues
Since it ends-up including `process.env`, it would change depending on how it
was built: `npm run demo-bundle`, `npm start`, or `npm run test:e2e:*`

Problem is, if the last way it was built was for running tests, the "in iframe"
demo would get disabled and wouldn't work on the demo page because that demo's
script isn't loaded in Cypress to prevent errors related to Cypress' lack of
iframe support.

So now, when building the bundle for testing with Cypress, the bundle gets
built as a new git-ignored `/docs/demo-bundle-cypress.*` set of files, and
building it for the docs page or for local debugging purposes builds it
as the normal `/docs/demo-bundle.js`.

When tests run and Cypress loads `/docs/index.html` it does so now by giving
it a search parameter, `/docs/index.html?bundle=cypress`, which results in
the `/docs/demo-bundle-cypress.js` script being loaded instead of the
`/docs/demo-bundle.js` non-Cypress-friendly script.

And since the `IS_CYPRESS_ENV:''` setting was removed from package.json when
running `npm start`, the `/docs/demo-bundle.js` shouldn't keep changing
solely based on env parameters since the same params (that affect the bundle's
contents) are used whether just building the bundle, or running it locally.