Refactor mutation observer usage
The current implementation is tied to state and event listeners.
Mutation observers does not quite fit here. This also leads to code
duplication.

1) The mutation observer is tied to container elements while event
listeners are tied to the document.
2) The mutation observer in itself cannot be considered state.

To address this the handling is moved to a separate section with a 
single observer (if supported) per trap and a single function
updateObservedNodes to reset the trap according to the current state
(active, pause and containers).

Call updateObservedNodes at all appropriate locations where trap state 
changes.

It would be more better if the mutationObserver could be declared with
other constants in createFocusTrap() but creating a new MutationObserver
instance requires a reference to the callback, checkDomRemoval(), which
must be declared before referenced.