Update lock file after recently publishing v7.4.2
This is the second time this happens. Seems to be something that has
stopped working WRT to the Changeset Bot because this is new, and
annoying behavior. The Bot should update the lock file when a
new release happens.

Or it's an NPM thing that it wants to update the lock file to the
current version, which innevitably changes upon publishing, but
only after publishing...