Ensure focus is kept inside trap when focussed element is removed (#932)
* Initial test setup when focused element in trap is removed

If the element that is currently focused is removed from the DOM, the
focus is moved either to the document's body element or becomes null
Source: https://amberwilson.co.uk/blog/where-did-the-focus-go/

When using focus trap it would be appropriate to keep focus inside the
trap.

This adds a Cypress test to show whether this works or not.

* Update DOM removal test check for undefined before checking for text

This avoids unwanted undefined errors when running tests.

* Pull focus back into DOM when focused element is removed

If the currently focused element is removed then browsers will reset
focus to window.body. See
https://allyjs.io/tutorials/mutating-active-element.html

This can cause assistive technologies to start from the beginning of a
document even a focus trap is set.

Use MutationObserver (if available) to monitor trap content. If the most
recent element to receive focus is removed then reset focus to the 
initial element in the trap.

* Only check for MutationObserver support in a client-side context

If running server-side window will be undefined.

* Use tabbable.isFocusable() to ensure that the initial node is focusable

This performs a set of checks including whether the node is connected to
the DOM.

* Reset mutation observer when container elements are updated

* Refactor mutation observer usage

The current implementation is tied to state and event listeners.
Mutation observers does not quite fit here. This also leads to code
duplication.

1) The mutation observer is tied to container elements while event
listeners are tied to the document.
2) The mutation observer in itself cannot be considered state.

To address this the handling is moved to a separate section with a 
single observer (if supported) per trap and a single function
updateObservedNodes to reset the trap according to the current state
(active, pause and containers).

Call updateObservedNodes at all appropriate locations where trap state 
changes.

It would be more better if the mutationObserver could be declared with
other constants in createFocusTrap() but creating a new MutationObserver
instance requires a reference to the callback, checkDomRemoval(), which
must be declared before referenced.

* Update demo bundle with with latest version

* Tweak /docs/index.js require order; update demo bundle again

---------

Co-authored-by: Stefan Cameron <stefan@stefcameron.com>